% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_solution.R
\name{b_solution_Results}
\alias{b_solution_Results}
\title{b_solution_Results}
\usage{
b_solution_Results(T1, T2, T3)
}
\arguments{
\item{T1}{Temp}

\item{T2}{Temp}

\item{T3}{Temp}
}
\value{
b_solution_Results
}
\description{
b_solution_Results
}
