% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{Calculate}
\alias{Calculate}
\title{Calculate}
\usage{
Calculate(compound, CP298, S298)
}
\arguments{
\item{compound}{A character string. The number of atoms will be automatically determined}

\item{CP298}{A numerical variable. The Heat capacity (Cp) value at the room temperature (298.15 K) in (j/mol K)}

\item{S298}{A numerical variable. The Entropy (S) value at the room temperature (298.15 K) in (j/mol K)}
}
\value{
The function returns the (Model outputs), (Model parameters) and
(polynomial_function_parameters) as 3 individual matrices
}
\description{
Calculate
}
