% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{calculate_S}
\alias{calculate_S}
\title{calculate_S}
\usage{
calculate_S(Temp, compound, CP298, S298)
}
\arguments{
\item{Temp}{Temperature range}

\item{compound}{A character string. The number of atoms will be automatically determined}

\item{CP298}{A numerical variable. The Heat capacity (Cp) value at the room temperature (298.15 K) in (j/mol K)}

\item{S298}{A numerical variable. The Entropy (S) value at the room temperature (298.15 K) in (j/mol K)}
}
\value{
the model's evaluation of Entropy (S)
}
\description{
calculate_S
}
