% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{Calphad_2.5}
\alias{Calphad_2.5}
\title{Calphad_2.5}
\usage{
Calphad_2.5(ele, CP100, CP200, CP298, S100, S200, S298)
}
\arguments{
\item{ele}{input parameter}

\item{CP100}{input parameter}

\item{CP200}{input parameter}

\item{CP298}{input parameter}

\item{S100}{input parameter}

\item{S200}{input parameter}

\item{S298}{input parameter}
}
\value{
Result
}
\description{
Calphad_2.5
}
