% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Debye_constant}
\alias{Debye_constant}
\title{Debye model}
\usage{
Debye_constant(x, Td)
}
\arguments{
\item{x}{temperature}

\item{Td}{Debye temperature}
}
\value{
Debye model
}
\description{
Debye model
}
\examples{
Debye_constant(500,250)
}
