% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Debye_constant_b}
\alias{Debye_constant_b}
\title{Debye model - with linear term}
\usage{
Debye_constant_b(x, Td, b)
}
\arguments{
\item{x}{temperature}

\item{Td}{Debye temperature}

\item{b}{linear term}
}
\value{
Debye model - with linear term
}
\description{
Debye model - with linear term
}
