% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Debye_constant_DOT}
\alias{Debye_constant_DOT}
\title{Debye model divided by T}
\usage{
Debye_constant_DOT(x)
}
\arguments{
\item{x}{Temperature}
}
\value{
Debye model divided by T
}
\description{
Debye model divided by T
}
