% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Debye_constant_DOT_b}
\alias{Debye_constant_DOT_b}
\title{Debye model divided by T - with linear term}
\usage{
Debye_constant_DOT_b(x)
}
\arguments{
\item{x}{Temperature}
}
\value{
Debye model divided by T - with linear term
}
\description{
Debye model divided by T - with linear term
}
