% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Debye_model_above}
\alias{Debye_model_above}
\title{Debye_model Temp dependent solution}
\usage{
Debye_model_above(x)
}
\arguments{
\item{x}{Temperature}
}
\value{
Debye_model Temp dependent solution
}
\description{
Debye_model Temp dependent solution
}
