% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Debye_model_below}
\alias{Debye_model_below}
\title{Debye_model_b_solution}
\usage{
Debye_model_below(x)
}
\arguments{
\item{x}{Temp}
}
\value{
Debye_model_b_solution
}
\description{
Debye_model_b_solution
}
