% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DebyeVarLin_1}
\alias{DebyeVarLin_1}
\title{Temp dependent debye model}
\usage{
DebyeVarLin_1(x, a)
}
\arguments{
\item{x}{Temperature}

\item{a}{constant parameter}
}
\value{
Temp dependent debye model
}
\description{
Temp dependent debye model
}
