% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DebyeVarLinOT_1}
\alias{DebyeVarLinOT_1}
\title{Temp dependent debye model divided by T. used to find the entropy}
\usage{
DebyeVarLinOT_1(x, a)
}
\arguments{
\item{x}{Temperature}

\item{a}{constant parameter}
}
\value{
Temp dependent debye model divided by T. used to find the entropy
}
\description{
Temp dependent debye model divided by T. used to find the entropy
}
