% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{entr_1_above}
\alias{entr_1_above}
\title{polynomial function for the first section of the fitted debye model (0-40) divided by T - Entropy}
\usage{
entr_1_above(x)
}
\arguments{
\item{x}{Temperature}
}
\value{
polynomial function for the first section of the fitted debye model (0-40) divided by T - Entropy
}
\description{
polynomial function for the first section of the fitted debye model (0-40) divided by T - Entropy
}
