% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{entr_1_below}
\alias{entr_1_below}
\title{polynomial function for the first section of the fitted debye model (0-40) divided by T}
\usage{
entr_1_below(x)
}
\arguments{
\item{x}{Temp}
}
\value{
polynomial function for the first section of the fitted debye model (0-40) divided by T
}
\description{
polynomial function for the first section of the fitted debye model (0-40) divided by T
}
