% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{entr_2_above}
\alias{entr_2_above}
\title{polynomial function for the second section of the fitted debye model (40-298.15) divided by T - Entropy}
\usage{
entr_2_above(x)
}
\arguments{
\item{x}{Temperature}
}
\value{
polynomial function for the second section of the fitted debye model (40-298.15) divided by T - Entropy
}
\description{
polynomial function for the second section of the fitted debye model (40-298.15) divided by T - Entropy
}
