% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{entr_2_below}
\alias{entr_2_below}
\title{polynomial function for the second section of the fitted debye model (40-298.15) divided by T}
\usage{
entr_2_below(x)
}
\arguments{
\item{x}{Temp}
}
\value{
polynomial function for the second section of the fitted debye model (40-298.15) divided by T
}
\description{
polynomial function for the second section of the fitted debye model (40-298.15) divided by T
}
