% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temp_Dep_solution.R
\name{g}
\alias{g}
\title{Find root for Td - Temp. Dep. Solution}
\usage{
g(Td)
}
\arguments{
\item{Td}{Debye temp}
}
\value{
Find root for Td - Temp. Dep. Solution
}
\description{
Find root for Td - Temp. Dep. Solution
}
