% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temp_Dep_solution.R
\name{Optimize_TDep_soution}
\alias{Optimize_TDep_soution}
\title{Optimize_TDep_soution}
\usage{
Optimize_TDep_soution(init)
}
\arguments{
\item{init}{initial value to calculate parameter a1}
}
\value{
calculate parameters a1 and b1 for Temp. Dep. solution
}
\description{
Optimize_TDep_soution
}
