% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{polyFunc_512pm}
\alias{polyFunc_512pm}
\title{5th order polynomial with x^-1 and x^-2}
\usage{
polyFunc_512pm(x, a, b, c, d, e, f, g)
}
\arguments{
\item{x}{Temperature}

\item{a}{fitting parameter}

\item{b}{fitting parameter}

\item{c}{fitting parameter}

\item{d}{fitting parameter}

\item{e}{fitting parameter}

\item{f}{fitting parameter}

\item{g}{fitting parameter}
}
\value{
5th order polynomial with x^-1 and x^-2
}
\description{
5th order polynomial with x^-1 and x^-2
}
