% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{polyFunc_6}
\alias{polyFunc_6}
\title{6th order polynomial}
\usage{
polyFunc_6(x, a, b, c, d, e, f, g)
}
\arguments{
\item{x}{Temperature}

\item{a}{fitting parameter}

\item{b}{fitting parameter}

\item{c}{fitting parameter}

\item{d}{fitting parameter}

\item{e}{fitting parameter}

\item{f}{fitting parameter}

\item{g}{fitting parameter}
}
\value{
6th order polynomial
}
\description{
6th order polynomial
}
