% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge_regression_b_solution.R
\name{Ridge_b_Solution}
\alias{Ridge_b_Solution}
\title{Ridge_b_Solution}
\usage{
Ridge_b_Solution(Td_input)
}
\arguments{
\item{Td_input}{parameter}
}
\value{
Ridge_b_Solution
}
\description{
Ridge_b_Solution
}
