% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ridge_regression_T_dep.R
\name{Ridge_T_Dep_Solution}
\alias{Ridge_T_Dep_Solution}
\title{Ridge_T_Dep_Solution}
\usage{
Ridge_T_Dep_Solution(a_input)
}
\arguments{
\item{a_input}{parameter}
}
\value{
Ridge_T_Dep_Solution
}
\description{
Ridge_T_Dep_Solution
}
