% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{S1_above}
\alias{S1_above}
\title{polynomial function for the first section of the fitted debye model (0-40)}
\usage{
S1_above(x)
}
\arguments{
\item{x}{#' @param x Temperature}
}
\value{
polynomial function for the first section of the fitted debye model (0-40)
}
\description{
polynomial function for the first section of the fitted debye model (0-40)
}
