% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{S1_below}
\alias{S1_below}
\title{polynomial function for the first section of the fitted debye model (0-40)}
\usage{
S1_below(x)
}
\arguments{
\item{x}{Temp}
}
\value{
polynomial function for the first section of the fitted debye model (0-40)
}
\description{
polynomial function for the first section of the fitted debye model (0-40)
}
