% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{S2_below}
\alias{S2_below}
\title{polynomial function for the second section of the fitted debye model (40-298.15)}
\usage{
S2_below(x)
}
\arguments{
\item{x}{Temp}
}
\value{
polynomial function for the second section of the fitted debye model (40-298.15)
}
\description{
polynomial function for the second section of the fitted debye model (40-298.15)
}
