% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temp_Dep_solution.R
\name{sum_square_error}
\alias{sum_square_error}
\title{sum_square_error}
\usage{
sum_square_error(a)
}
\arguments{
\item{a}{parameter}
}
\value{
returns the sum square error
}
\description{
sum_square_error
}
