% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temp_Dep_solution.R
\name{Temp_Dep_Results}
\alias{Temp_Dep_Results}
\title{Temp_Dep_Results}
\usage{
Temp_Dep_Results(T1, T2, T3)
}
\arguments{
\item{T1}{Temp}

\item{T2}{Temp}

\item{T3}{Temp}
}
\value{
Temp_Dep_Results
}
\description{
Temp_Dep_Results
}
